requirements.txt
anthropic>=0.30.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
selenium>=4.18.0
webdriver-manager>=4.0.0
yt-dlp>=2024.4.9
